// backend/src/types/studySession.ts
export enum StudyStatus {
  Studying = 'STUDYING',
  Break = 'BREAK',
  Finished = 'FINISHED',
}

export enum StudyState {
  Focus = 'FOCUS',
  Distracted = 'DISTRACTED',
  Break = 'BREAK',
}

export enum DistractedSubtype {
  Phone = 'PHONE',
  Sleep = 'SLEEP',
  Leave = 'LEAVE',
  LookAround = 'LOOK_AROUND',
  HeadDown = 'HEAD_DOWN',
  Other = 'OTHER',
}